Backend/ocr-module/sign/requirements.txt
flask
flask_sqlalchemy
Pillow
opencv-python
tensorflow
Flask
pytesseract
# Any other dependencies
